@@ -1,8 +1,9 @@
-package com.andalus.abomedhat.mohamedfawzytasks;
+package com.example.android.weather;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+
+import androidx.test.InstrumentationRegistry;
+import androidx.test.runner.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -21,6 +22,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getTargetContext();
 
-        assertEquals("com.andalus.abomedhat.mohamedfawzytasks", appContext.getPackageName());
+        assertEquals("com.example.android.weather", appContext.getPackageName());
     }
 }
